Test 'config' package renamed as 'configuration' to match the src folder, VariantJobsArgs moved to 'uk.ac.ebi.eva.pipeline.configuration'
