+++ src/main/java/uk/ac/ebi/eva/pipeline/configuration/VariantJobsArgs.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva;
+package uk.ac.ebi.eva.pipeline.configuration;
 
 import java.io.FileInputStream;
 import java.io.IOException;
