+++ src/main/java/uk/ac/ebi/eva/pipeline/jobs/steps/tasklet/VariantsStatsCreate.java
@@ -33,7 +33,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
 import org.springframework.stereotype.Component;
 
-import uk.ac.ebi.eva.VariantJobsArgs;
+import uk.ac.ebi.eva.pipeline.configuration.VariantJobsArgs;
 import uk.ac.ebi.eva.utils.URLHelper;
 
 import java.net.URI;
