+++ src/test/java/uk/ac/ebi/eva/pipeline/configuration/AnnotationConfig.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *          http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.ac.ebi.eva.pipeline.config;
+package uk.ac.ebi.eva.pipeline.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -24,11 +24,11 @@ import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
 /**
  * @author Diego Poggioli
  *
- * Configuration init for VariantConfiguration Job
+ * Configuration init for Annotation Job
  */
 @Configuration
-@PropertySource({"variantConfiguration.properties"})
-public class VariantConfig {
+@PropertySource({"annotation.properties"})
+public class AnnotationConfig {
 
     @Bean
     private static PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer() {
