+++ src/test/java/uk/ac/ebi/eva/pipeline/jobs/VariantAggregatedConfigurationTest.java
@@ -15,8 +15,8 @@
  */
 package uk.ac.ebi.eva.pipeline.jobs;
 
-import uk.ac.ebi.eva.VariantJobsArgs;
-import uk.ac.ebi.eva.pipeline.config.CommonConfig;
+import uk.ac.ebi.eva.pipeline.configuration.CommonConfig;
+import uk.ac.ebi.eva.pipeline.configuration.VariantJobsArgs;
 import uk.ac.ebi.eva.pipeline.jobs.VariantAggregatedConfiguration;
 
 import org.junit.AfterClass;
