+++ src/test/java/uk/ac/ebi/eva/pipeline/jobs/VariantConfigurationTest.java
@@ -52,8 +52,8 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.zip.GZIPInputStream;
 
-import uk.ac.ebi.eva.VariantJobsArgs;
-import uk.ac.ebi.eva.pipeline.config.VariantConfig;
+import uk.ac.ebi.eva.pipeline.configuration.VariantConfig;
+import uk.ac.ebi.eva.pipeline.configuration.VariantJobsArgs;
 import uk.ac.ebi.eva.pipeline.jobs.steps.VariantsAnnotLoad;
 
 import static junit.framework.TestCase.assertEquals;
