+++ src/test/java/uk/ac/ebi/eva/pipeline/jobs/VariantConfigurationWorkflowTest.java
@@ -16,8 +16,8 @@
 
 package uk.ac.ebi.eva.pipeline.jobs;
 
-import uk.ac.ebi.eva.VariantJobsArgs;
-import uk.ac.ebi.eva.pipeline.config.VariantWorkflowConfig;
+import uk.ac.ebi.eva.pipeline.configuration.VariantJobsArgs;
+import uk.ac.ebi.eva.pipeline.configuration.VariantWorkflowConfig;
 import uk.ac.ebi.eva.pipeline.jobs.steps.VariantsAnnotGenerateInput;
 import uk.ac.ebi.eva.pipeline.jobs.steps.VariantsAnnotLoad;
 import uk.ac.ebi.eva.pipeline.jobs.steps.tasklet.VariantsAnnotCreate;
