+++ src/test/java/uk/ac/ebi/eva/pipeline/jobs/steps/VariantsAnnotGenerateInputTest.java
@@ -18,9 +18,9 @@ package uk.ac.ebi.eva.pipeline.jobs.steps;
 import com.mongodb.*;
 import com.mongodb.util.JSON;
 
-import uk.ac.ebi.eva.VariantJobsArgs;
 import uk.ac.ebi.eva.pipeline.annotation.generateInput.VariantAnnotationItemProcessor;
-import uk.ac.ebi.eva.pipeline.config.AnnotationConfig;
+import uk.ac.ebi.eva.pipeline.configuration.AnnotationConfig;
+import uk.ac.ebi.eva.pipeline.configuration.VariantJobsArgs;
 import uk.ac.ebi.eva.pipeline.io.readers.VariantReader;
 import uk.ac.ebi.eva.pipeline.io.writers.VepInputWriter;
 import uk.ac.ebi.eva.pipeline.jobs.JobTestUtils;
